requirements.txt
Click==7.0
Flask==1.1.1
Flask-Login==0.5.0	
Flask-MySQLdb==0.2.0	
Flask-SQLAlchemy==2.4.1	
Jinja2==2.11.1	
Keras==2.2.4	
Keras-Applications==1.0.8	
Keras-Preprocessing==1.1.0	
Markdown==3.1.1	
MarkupSafe==1.1.1	
Pillow==6.1.0	
PyAudio==0.2.11	
PyNaCl==1.3.0	
PyYAML==5.1.2	
SQLAlchemy==1.3.15
SoundFile==0.10.3
Werkzeug==0.15.5
absl-py==0.7.1
astor==0.8.0
attrdict==2.0.1
audioread==2.1.8
bcrypt==3.1.7
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
cycler==0.10.0
decorator==4.4.1
gast==0.2.2
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.9.0
idna==2.8
itsdangerous==1.1.0
joblib==0.14.1
kiwisolver==1.1.0
librosa==0.7.2
llvmlite==0.31.0
lomond==0.3.3
matplotlib==3.1.2
mysqlclient==1.4.6
numba==0.47.0
numpy==1.18.1
opt-einsum==3.1.0
pandas==0.25.3
paramiko==2.7.1
pip==19.2.1
progressbar2==3.47.0
protobuf==3.9.1
pycparser==2.19
pyparsing==2.4.6
python-dateutil==2.8.1
python-utils==2.3.0
pytz==2019.3
pyxdg==0.26
requests==2.22.0
resampy==0.2.2
scikit-learn==0.22.1
scipy==1.3.0
setuptools==40.8.0
six==1.12.0
soupsieve==1.9.5
sox==1.3.7
termcolor==1.1.0
urllib3==1.25.7
virtualenv==16.7.9
webrtcvad==2.0.10
wheel==0.33.4
wrapt==1.11.2
pillow==7.1.1